+++ changelog.md
@@ -18,7 +18,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 ## Fixed
 
-- Prevent equal-crossing binary operands from producing overlapping connectors (``)
+- Prevent equal-crossing binary operands from producing overlapping connectors (`5507e7c`)
+- Remove conflict between `EdgeAnchors` and binary operator input routing (``)
 
 # [Released]